Detect MTP from the source in TestEngine, not ExecutionPreference
TestEngine already has the source (SourceDetail.Source) at every point where
it decides how to route a run, so let it derive whether a source is a
Microsoft.Testing.Platform app from the assembly itself instead of reading a
pre-baked ExecutionPreference that had to be computed upstream and threaded
through SourceDetail/RunConfiguration.

The MTP detection (the [assembly: AssemblyMetadata(...)] PEReader probe) moves
into a shared MicrosoftTestingPlatformDetector in CoreUtilities, which already
uses PEReader and is befriended by CrossPlatEngine, vstest.console and the
TestHostRuntimeProvider. vstest.console's AssemblyMetadataProvider now delegates
to it instead of carrying its own copy of the (subtle) attribute-blob parsing.

TestEngine memoizes the result per source path (concurrent, since the parallel
manager creators run on multiple threads). ExecutionPreference is still set
upstream for now; TestEngine simply no longer depends on it. Grouping,
isolation, the sentinel provider and the MTP proxy routing all key on the
source instead.

Verified: pure-MTP asset still runs Failed1/Passed2/Skipped1/Total4 under
vstest.console; CrossPlatEngine.UnitTests (665) and Common.UnitTests (396) green.
